# contracts/services-openAPI.yaml
openapi: 3.0.3
info:
  title: Amazon Order Processing Services API
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /billing/accounts/{accountId}:
    get:
      tags: [Billing]
      summary: Get billing account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Billing account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingAccount'
...
  /billing/refunds:
    post:
      tags: [Billing]
      summary: Refund a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
...
components:
...
  schemas:
...
    BillingAccount:
      type: object
      properties:
        account_id:
          type: string
          format: uuid
        balance:
          type: number
...
    RefundRequest:
      type: object
      required: [order_id, amount]
      properties:
        order_id:
          type: string
          format: uuid
        amount:
          type: number
    RefundResponse:
      type: object
      properties:
        refund_id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        amount:
          type: number
...